+++ hookee/cli.py
@@ -114,9 +114,7 @@ def enable_plugin(ctx, plugin):
     """
     cli_manager = ctx.obj["cli_manager"]
 
-    loaded_plugin = cli_manager.plugin_manager.get_plugin(plugin)
-
-    cli_manager.plugin_manager.validate_plugin(loaded_plugin)
+    cli_manager.plugin_manager.get_plugin(plugin)
 
     cli_manager.config.append("plugins", plugin)